requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
Brotli==1.0.9
cffi==1.15.0
click==8.0.4
cssselect2==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-easy-pdf==0.1.1
django-utils-six==2.0
djangorestframework==3.13.1
fonttools==4.29.1
future==0.18.2
html5lib==1.1
mypy-extensions==0.4.3
pathspec==0.9.0
pdfkit==1.0.0
Pillow==9.0.1
platformdirs==2.5.1
pycparser==2.21
pydyf==0.1.2
PyPDF2==1.26.0
pyphen==0.12.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tomli==2.0.1
typing-extensions==4.1.1
weasyprint==54.2
webencodings==0.5.1
xhtml2pdf==0.2.5
zopfli==0.2.1
